+++ service_crategen/src/service.rs
@@ -111,7 +111,7 @@ impl <'b> Service <'b> {
                 dependencies.insert("serde_json".to_owned(), cargo::Dependency::Simple("1.0.1".into()));
             },
             "query" | "ec2" | "rest-xml" => {
-                dependencies.insert("xml-rs".to_owned(), cargo::Dependency::Simple("0.6".into()));
+                dependencies.insert("xml-rs".to_owned(), cargo::Dependency::Simple("0.7".into()));
             },
             "rest-json" => {
                 dependencies.insert("log".to_owned(), cargo::Dependency::Simple("0.3.6".into()));