+++ CHANGELOG.md
@@ -19,8 +19,10 @@
 - Added Elastic Filesystem service
 - Fix broken links on the STS documentation
 - Use xml::EventWriter to serialize XML payload
+- Added Workdocs service
 - Added Cognito Sync service
 - Added Glue service
+- Added DynamoDB Accelerator
 
 ## [0.28.0] - 2017-08-25